feat: base and penalty tests
